third_party/blink/web_tests/editing/selection/select_all/select_all_details_crash.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../assert_selection.js"></script>
<div id="log"></div>
<script>
test(() => {
    assert_selection(
        [
            '|<legend style="all: initial; -webkit-user-modify: initial">foo</legend>',
            '<details style="display: inline"></details>',
            '<textarea></textarea>',
        ].join(''),
        selection => {
            selection.document.execCommand('selectAll');
            selection.document.designMode = 'on';
        },
        [
            '<legend style="all: initial; -webkit-user-modify: initial">^foo|</legend>',
            '<details style="display: inline"></details>',
            '<textarea></textarea>',
        ].join(''));
}, 'selectAll on DETAILS should not crash');
</script>
